Fix the deprecation message for inheritance deprecation.
Move to a 2-object collaborative design to do so, where
(Emitted)Deprecations delegate to more specific deprecation
representatives.

This is going to come in handy for parameter deprecation
too.

Also fills in the missing integration test that would have
caught this.

There's still plenty of room to make this smoother, it's still
a bit clunky, but it's at least better than what was here in
terms of flexibility.