Fix more wrong stack levels.
Rest of the fallout from fa2cee6 and / or
https://twistedmatrix.com/trac/ticket/9363.

Here, __file__ *is* correct, because we're
wrapping the inheritance in a function, so it
*will* in fact happen in the test_integration
file and not in assertWarns' defining module.

Hooray...

(I guess I'm at least happy to kill the extra
complexity around deprecations being able to
add extra stack levels to remove. It's not
needed, at least not yet.)